# Optimize OnHeapHnswGraph's data structure