--- a/lucene/core/src/java/org/apache/lucene/util/hnsw/OnHeapHnswGraph.java
+++ b/lucene/core/src/java/org/apache/lucene/util/hnsw/OnHeapHnswGraph.java
@@ -20,10 +20,9 @@ package org.apache.lucene.util.hnsw;
 import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import org.apache.lucene.util.Accountable;
+import org.apache.lucene.util.ArrayUtil;
 import org.apache.lucene.util.RamUsageEstimator;
 
 /**
@@ -31,6 +30,8 @@ import org.apache.lucene.util.RamUsageEstimator;
  * construct the HNSW graph before it's written to the index.
  */
 public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
+
+  private static final int INIT_SIZE = 128;
 
   private int numLevels; // the current number of levels in the graph
   private int entryNode; // the current graph entry node on the top level. -1 if not set
@@ -40,31 +41,29 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
   // to vectors
   // added to HnswBuilder, and the node values are the ordinals of those vectors.
   // Thus, on all levels, neighbors expressed as the level 0's nodes' ordinals.
-  private final List<NeighborArray> graphLevel0;
-  // Represents levels 1-N. Each level is represented with a Map that maps a levels level 0
-  // ordinal to its neighbors on that level. All nodes are in level 0, so we do not need to maintain
-  // it in this list. However, to avoid changing list indexing, we always will make the first
-  // element
-  // null.
-  private final List<Map<Integer, NeighborArray>> graphUpperLevels;
-  private final int nsize;
-  private final int nsize0;
+  private NeighborArray[][] graph;
+  private  List<Integer>[] levelToNodes;
+  private int lastFreezeSize;
+  private int size; // graph size, which is number of nodes in level 0
+  private int nonZeroLevelSize; // total number of NeighborArrays created that is not on level 0
+  private final int nsize; // neighbour array size at non-zero level
+  private final int nsize0; // neighbour array size at zero level
 
   // KnnGraphValues iterator members
   private int upto;
   private NeighborArray cur;
 
-  OnHeapHnswGraph(int M) {
+  OnHeapHnswGraph(int M, int numNodes) {
     this.numLevels = 1; // Implicitly start the graph with a single level
-    this.graphLevel0 = new ArrayList<>();
     this.entryNode = -1; // Entry node should be negative until a node is added
     // Neighbours' size on upper levels (nsize) and level 0 (nsize0)
     // We allocate extra space for neighbours, but then prune them to keep allowed maximum
     this.nsize = M + 1;
     this.nsize0 = (M * 2 + 1);
-
-    this.graphUpperLevels = new ArrayList<>(numLevels);
-    graphUpperLevels.add(null); // we don't need this for 0th level, as it contains all nodes
+    if (numNodes == -1) {
+      numNodes = INIT_SIZE;
+    }
+    this.graph = new NeighborArray[numNodes][];
   }
 
   /**
@@ -74,22 +73,20 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
    * @param node the node whose neighbors are returned, represented as an ordinal on the level 0.
    */
   public NeighborArray getNeighbors(int level, int node) {
-    if (level == 0) {
-      return graphLevel0.get(node);
-    }
-    Map<Integer, NeighborArray> levelMap = graphUpperLevels.get(level);
-    assert levelMap.containsKey(node);
-    return levelMap.get(node);
+    assert graph.length > node && graph[node] != null && graph[node][level] != null;
+    return graph[node][level];
   }
 
   @Override
   public int size() {
-    return graphLevel0.size(); // all nodes are located on the 0th level
+    return size;
   }
 
   /**
    * Add node on the given level. Nodes can be inserted out of order, but it requires that the nodes
    * preceded by the node inserted out of order are eventually added.
+   *
+   * NOTE: You must add a node from the node's top level
    *
    * @param level level to add a node on
    * @param node the node to add, represented as an ordinal on the level 0.
@@ -99,27 +96,25 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
       entryNode = node;
     }
 
-    if (level > 0) {
-      // if the new node introduces a new level, add more levels to the graph,
-      // and make this node the graph's new entry point
-      if (level >= numLevels) {
-        for (int i = numLevels; i <= level; i++) {
-          graphUpperLevels.add(new HashMap<>());
-        }
-        numLevels = level + 1;
-        entryNode = node;
-      }
+    if (node >= graph.length) {
+      graph = ArrayUtil.grow(graph, node + 1);
+    }
 
-      graphUpperLevels.get(level).put(node, new NeighborArray(nsize, true));
+    if (level >= numLevels) {
+      numLevels = level + 1;
+      entryNode = node;
+    }
+
+    assert graph[node] == null || graph[node].length > level;
+    if (graph[node] == null) {
+      graph[node] = new NeighborArray[level + 1]; // assumption: we always call this function from top level
+      size++;
+    }
+    if (level == 0) {
+      graph[node][level] = new NeighborArray(nsize0, true);
     } else {
-      // Add nodes all the way up to and including "node" in the new graph on level 0. This will
-      // cause the size of the
-      // graph to differ from the number of nodes added to the graph. The size of the graph and the
-      // number of nodes
-      // added will only be in sync once all nodes from 0...last_node are added into the graph.
-      while (node >= graphLevel0.size()) {
-        graphLevel0.add(new NeighborArray(nsize0, true));
-      }
+      graph[node][level] = new NeighborArray(nsize, true);
+      nonZeroLevelSize++;
     }
   }
 
@@ -163,8 +158,36 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
     if (level == 0) {
       return new ArrayNodesIterator(size());
     } else {
-      return new CollectionNodesIterator(graphUpperLevels.get(level).keySet());
+      generateLevelToNodes();
+      return new CollectionNodesIterator(levelToNodes[level]);
     }
+  }
+
+  @SuppressWarnings({"unchecked", "rawtypes"})
+  private void generateLevelToNodes() {
+    if (lastFreezeSize == size) {
+      return;
+    }
+
+    levelToNodes = new List[numLevels];
+    for (int i = 1; i < numLevels; i++) {
+      levelToNodes[i] = new ArrayList<>();
+    }
+    int nonNullNode = 0;
+    for (int node = 0; node < graph.length; node++) {
+      // when we init from another graph, we could have holes where some slot is null
+      if (graph[node] == null) {
+        continue;
+      }
+      nonNullNode++;
+      for (int i = 1; i < graph[node].length; i++) {
+        levelToNodes[i].add(node);
+      }
+      if (nonNullNode == size) {
+        break;
+      }
+    }
+    lastFreezeSize = size;
   }
 
   @Override
@@ -180,29 +203,9 @@ public final class OnHeapHnswGraph extends HnswGraph implements Accountable {
             + RamUsageEstimator.NUM_BYTES_OBJECT_REF * 2
             + Integer.BYTES * 3;
     long total = 0;
-    for (int l = 0; l < numLevels; l++) {
-      if (l == 0) {
-        total +=
-            graphLevel0.size() * neighborArrayBytes0
-                + RamUsageEstimator.NUM_BYTES_OBJECT_REF; // for graph;
-      } else {
-        long numNodesOnLevel = graphUpperLevels.get(l).size();
-
-        // For levels > 0, we represent the graph structure with a tree map.
-        // A single node in the tree contains 3 references (left root, right root, value) as well
-        // as an Integer for the key and 1 extra byte for the color of the node (this is actually 1
-        // bit, but
-        // because we do not have that granularity, we set to 1 byte). In addition, we include 1
-        // more reference for
-        // the tree map itself.
-        total +=
-            numNodesOnLevel * (3L * RamUsageEstimator.NUM_BYTES_OBJECT_REF + Integer.BYTES + 1)
-                + RamUsageEstimator.NUM_BYTES_OBJECT_REF;
-
-        // Add the size neighbor of each node
-        total += numNodesOnLevel * neighborArrayBytes;
-      }
-    }
+    total += size * (neighborArrayBytes0 + RamUsageEstimator.NUM_BYTES_ARRAY_HEADER)
+                    + RamUsageEstimator.NUM_BYTES_ARRAY_HEADER; // for graph and level 0;
+    total += nonZeroLevelSize * neighborArrayBytes;
     return total;
   }