… lucene/CHANGES.txt
… lucene/core/src/test/org/apache/lucene/util/hnsw/HnswGraphTestCase.java
… lucene/core/src/test/org/apache/lucene/util/hnsw/TestOnHeapHnswGraph.java