### Description

Make the `OnHeapHnswGraph` essentially a 2D array of `NeighbourArray`, there're several benefits of this change:
1. We do less resizing of arrays: previously we're using `ArrayList` for level0 and `TreeMap` for rest of the levels which both have overheads on resizing or inserting. After this change, we don't need `TreeMap` anymore, and during indexing time, we only need to resize one array no matter how many levels we have, and the lookup will be a simpler 2d array lookup.
2. Multithread should be easier: this PR doesn't contain any attempt to make indexing concurrent, but in case that we want to make merging HNSW graph concurrent, this change makes things much easier, because when we merge we already know the total size of graph so the first dimension of the graph array is fixed, such that we don't need any sync on the graph for it to resize.

The only regression on this approach is that when we get node for a non-zero level it need to traverse the whole graph. But since that is not a common operation and is only called when we serialize to disk, I made a cache for it such that the cost of first non-zero level call is O(whole_graph) and subsequent call will be trivial.

### Test runs
I use writer buffer of 256MB and forcemerge at the end, and measured forceMerge time as well
#### Baseline
```
recall	latency	nDoc	fanout	maxConn	beamWidth	visited	index ms
Force merge done, time: 3 ms
0.838	 0.28	10000	0	64	250	100	3868	1.00	post-filter
Force merge done, time: 2 ms
0.755	 1.21	100000	0	64	250	100	101113	1.00	post-filter
Force merge done, time: 632325 ms
0.603	10.27	1000000	0	64	250	100	1622348	1.00	post-filter
```
#### Candidate
```
recall	latency	nDoc	fanout	maxConn	beamWidth	visited	index ms
Force merge done, time: 2 ms
0.838	 0.27	10000	0	64	250	100	3652	1.00	post-filter
Force merge done, time: 2 ms
0.755	 1.13	100000	0	64	250	100	92449	1.00	post-filter
Force merge done, time: 625726 ms
0.608	10.22	1000000	0	64	250	100	1529417	1.00	post-filter
```