@msokolov 2023-10-12
I like this! Actually I think when we are merging we can preallocate the entire array so we don't need to resize at all which should greatly simplify making this beast thread-safe (since the array at least will be immutable).