@zhaih 2023-10-12
Yes that's the idea, although I actually made some mistakes here so the
merging is not entirely pre-allocated, also something in searching might be
broken due to the size() behavior change, I will try to fix them later on.

On Wed, Oct 11, 2023, 21:17 Michael Sokolov ***@***.***>
wrote:

> I like this! Actually I think when we are merging we can preallocate the
> entire array so we don't need to resize at all which should greatly
> simplify making this beast thread-safe (since the array at least will be
> immutable).
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/apache/lucene/pull/12651#issuecomment-1758885797>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AEFSB7FHP6Y6SIQTOU5H6BDX65VPPANCNFSM6AAAAAA525PM6Q>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>