@zhaih 2023-10-12
OK I have incorporate all the learning I have from #12660 and added several more assertions to make it safer, please take a look again when you have time @msokolov, thanks!