@zhaih 2023-10-16
I reran'd the benchmark and still get the similar perf and same recall. (Just to make sure the later edits have not messed up things)